fixed ai alert debug display so that it works
git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@648 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ game/ai/ai.cpp
@@ -7,6 +7,9 @@
  * $Author$
  *
  * $Log$
+ * Revision 1.37  2006/12/10 03:29:43  ishtvan
+ * fixed ai alert debug display so that it works
+ *
  * Revision 1.36  2006/12/10 02:55:15  ishtvan
  * *) attempted to make AI vision better
  *
@@ -1561,7 +1564,7 @@ void idAI::Think( void ) {
 
 	if( cv_ai_alertnum_show.GetBool() )
 	{
-		gameRenderWorld->DrawText( va("Alert: %f", AI_AlertNum), (GetEyePosition() - physicsObj.GetGravityNormal()*32.0f), 0.25f, colorWhite, gameLocal.GetLocalPlayer()->viewAngles.ToMat3(), 1, gameLocal.msec );
+		gameRenderWorld->DrawText( va("Alert: %f", (float) AI_AlertNum), (GetEyePosition() - physicsObj.GetGravityNormal()*32.0f), 0.25f, colorWhite, gameLocal.GetLocalPlayer()->viewAngles.ToMat3(), 1, gameLocal.msec );
 	}
 }
 
@@ -6088,7 +6091,7 @@ float idAI::getPlayerVisualStimulusAmount(idEntity* p_playerEntity)
 		// Old method, commented out
 		// alertAmount = 4*log( visFrac * lgem ) / 0.6931472;
 
-		float CurAlert = AI_AlertNum;
+		float CurAlert = (float) AI_AlertNum;
 		// convert current alert from log to linear scale, add, then convert back
 		// this might not be as good for performance, but it lets us keep all alerts
 		// on the same scale.
